requirements.txt
Django
djangorestframework
psycopg2-binary
python-dotenv
asgiref
sqlparse
django-phonenumber-field
phonenumberslite
pillow
django-filter
djangorestframework-simplejwt==5.3.1
